Add Blocks Reordering

+++ database/migrations/2023_10_04_113617_create_blocks_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->constrained()
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
-            $table->unsignedInteger('order');
+            $table->decimal('order');
             $table->foreignId('component_id')
                 ->nullable()
                 ->constrained()
